+++ CHANGES.rst
@@ -1,6 +1,7 @@
 4.0.0 (Unreleased)
 ==================
 
+- Fix large-cube memory issues with channel-wise processing (#323)
 - Added initial test suite, which checks the various CASA tasks and arguments used throughout the pipeline (#324)
 - Bump actions/upload-artifact from 6 to 7 (#313)
 - Bump casaplotms requirement from >=2.7.4 to >=2.8.2 (#320)